requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
appnope==0.1.4
astroid==3.2.2
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.6.2
charset-normalizer==3.3.2
comm==0.2.2
dataclasses-json==0.6.6
debugpy==1.8.1
decorator==5.1.1
dill==0.3.8
distro==1.9.0
exceptiongroup==1.2.1
executing==2.0.1
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
iniconfig==2.0.0
ipykernel==6.29.4
ipython==8.25.0
isort==5.13.2
jedi==0.19.1
jsonpatch==1.33
jsonpointer==2.4
jupyter_client==8.6.2
jupyter_core==5.7.2
langchain==0.2.2
langchain-community==0.2.2
langchain-core==0.2.4
langchain-openai==0.1.8
langchain-text-splitters==0.2.1
langsmith==0.1.71
marshmallow==3.21.2
matplotlib-inline==0.1.7
mccabe==0.7.0
multidict==6.0.5
mypy-extensions==1.0.0
neo4j==5.20.0
nest-asyncio==1.6.0
numpy==1.26.4
openai==1.30.5
orjson==3.10.3
packaging==23.2
parso==0.8.4
pexpect==4.9.0
platformdirs==4.2.2
pluggy==1.5.0
prompt_toolkit==3.0.46
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==2.7.2
pydantic_core==2.18.3
Pygments==2.18.0
pylint==3.2.2
PyPDF2==3.0.1
pytest==8.2.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
regex==2024.5.15
requests==2.32.3
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.30
stack-data==0.6.3
tenacity==8.3.0
tiktoken==0.7.0
tomli==2.0.1
tomlkit==0.12.5
tornado==6.4
tqdm==4.66.4
traitlets==5.14.3
typing-inspect==0.9.0
typing_extensions==4.12.1
urllib3==2.2.1
wcwidth==0.2.13
yarl==1.9.4
